requirements.txt
anyio==3.2.1
appdirs==1.4.4
argon2-cffi @ file:///D:/bld/argon2-cffi_1605217273648/work
astroid==2.5
async-generator==1.10
atomicwrites==1.4.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
Babel==2.9.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1612213472466/work
certifi==2021.10.8
cffi @ file:///D:/bld/cffi_1613414140225/work
chardet==4.0.0
charset-normalizer==2.0.12
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
datapungi-fed==0.3.1
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
distlib==0.3.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
et-xmlfile==1.1.0
filelock==3.0.12
h11==0.12.0
html5lib==1.1
httpcore==0.14.7
httpx==0.22.0
idna==2.10
importlib-metadata @ file:///D:/bld/importlib-metadata_1622775666625/work
iniconfig==1.1.1
investpy==1.0.8
ipykernel @ file:///D:/bld/ipykernel_1620913090465/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1622679407603/work
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.1
jedi @ file:///D:/bld/jedi_1610146948662/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
json5==0.9.6
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter==1.0.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1615693636836/work
jupyter-console==6.4.0
jupyter-core @ file:///D:/bld/jupyter_core_1612125450516/work
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.6.0
lxml==4.8.0
MarkupSafe @ file:///D:/bld/markupsafe_1621455847127/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
mccabe==0.6.1
mistune @ file:///D:/bld/mistune_1610113104461/work
multitasking==0.0.10
nbclassic==0.3.1
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1614336084111/work
nbconvert @ file:///D:/bld/nbconvert_1624472961560/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1621259862661/work
numpy @ file:///D:/bld/numpy_1651020737746/work
openpyxl==3.0.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas==1.2.5
pandocfilters==1.4.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pluggy==1.0.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1622586138406/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1623977816122/work
py==1.11.0
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1620245170812/work
pylint==2.6.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent @ file:///D:/bld/pyrsistent_1610147004705/work
pytest==7.1.2
pytest-html==3.1.1
pytest-metadata==2.0.1
python-bcb==0.1.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.1
PyYAML==6.0
pyzmq @ file:///D:/bld/pyzmq_1622038669039/work
qtconsole==5.1.0
QtPy==1.9.0
requests==2.27.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1624366715919/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.2.0
soupsieve==2.3.2.post1
TA-Lib @ file:///D:/bld/ta-lib_1637267093594/work
terminado==0.10.1
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
tornado @ file:///D:/bld/tornado_1610094858390/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
Unidecode==1.3.4
urllib3==1.26.5
virtualenv==20.2.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websocket-client==1.1.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
yfinance==0.1.70
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
